% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaValidate.R
\name{iaValidate}
\alias{iaValidate}
\title{Merges Variables Into a Date Variable}
\usage{
iaValidate(df, col, samplesize = 5, reset = FALSE, idcol = "id")
}
\arguments{
\item{df}{Dataset to work with}

\item{col}{Column containing data to be displayed}

\item{samplesize}{How many entries to display at once}

\item{reset}{Reset the process / all "seen" <- 0}

\item{idcol}{Values from idcol are displayed next to values from col.}
}
\description{
This function lets you go through data, labelling it as either
1 or 0. Useful for validating data. 
Enter numbers corresponding to entries displayed, unseparated.
}
\examples{
mtcars$names <- rownames(mtcars)
mtcars$brand <- lapply(strsplit(mtcars$names,' '),function(x)x[1]) \%>\% unlist()
valMt <- iaValidate(mtcars,col = 'names',samplesize = 5,idcol = 'brand') 
}
\keyword{validation}
